se hace marge de caso de usos 1 y 2
