--- src/main/java/co/tarjetaCredito/Controladores/ClienteNoEncontradoException.java
@@ -1,8 +1,0 @@
-package co.tarjetaCredito.Controladores;
-
-public class ClienteNoEncontradoException extends RuntimeException {
-    public ClienteNoEncontradoException(String message) {
-        super(message);
-    }
-}
-
