+++ src/main/java/co/tarjetaCredito/Entidades/Empleado.java
@@ -1,0 +1,31 @@
+package co.tarjetaCredito.entidades;
+
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.io.Serializable;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+@Entity
+@Table(name = "Empleado")
+public class Empleado implements Serializable {
+    @Id
+    @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "SEQ_EMPLEADO")
+    @SequenceGenerator(name = "SEQ_EMPLEADO", sequenceName = "SEQ_EMPLEADO", allocationSize = 1)
+    @Column(name = "EMP_CODIGO")
+    private long serial;
+
+    @Column(name = "EMP_NOMBRE")
+    private String nombre;
+
+    @Column(name = "EMP_CORREO")
+    private String correo;
+
+    @Column(name = "EMP_CONTRASENA")
+    private String contrasena;
+
+}
