--- src/main/java/co/tarjetaCredito/TarjetaCreditoApplication.java
@@ -1,23 +1,0 @@
-package co.tarjetaCredito;
-
-
-
-import org.springframework.boot.CommandLineRunner;
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-import java.util.List;
-
-/**
- * author Vinni 2023
- */
-@SpringBootApplication
-
-public class TarjetaCreditoApplication {
-    public static void main(String[] args) {
-        SpringApplication.run(TarjetaCreditoApplication.class, args);
-    }
-
-   
-
-}
