+++ src/main/java/co/tarjetaCredito/repositorios/ClienteRepo.java
@@ -2,11 +2,10 @@ package co.tarjetaCredito.repositorios;
 
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
-
-import co.tarjetaCredito.entidades.Cliente;
+import co.tarjetaCredito.Entidades.ClientesEntidad;
 
 @Repository
-public interface ClienteRepo extends CrudRepository<Cliente, Integer>{
+public interface ClienteRepo extends CrudRepository<ClientesEntidad, Integer>{
 
     
 }
