--- src/main/java/co/tarjetaCredito/servicios/ClienteServ.java
@@ -1,7 +1,0 @@
-package co.tarjetaCredito.servicios;
-
-import java.util.List;
-import co.tarjetaCredito.entidades.Cliente;
-public interface ClienteServ {
-    List<Cliente> obtenerClientes();
-}
