+++ src/main/resources/templates/cliente.html
@@ -87,7 +87,7 @@
             <input type="password" id="password" name="password" required>
             <input type="submit" value="Iniciar sesión">
         </form>
-        <p>¿No tienes una cuenta? <a href="/registro">Regístrate aquí</a></p>
+        <p>¿No tienes una cuenta? <a href="#" th:href="@{/registro}">Regístrate</a></p>
     </div>
 </body>
 </html>
