+++ src/main/resources/templates/index.html
@@ -61,11 +61,11 @@
     
     <script>
         function redirectToCliente() {
-            window.location.href = "/validarEmpleado";
+            window.location.href = "/validarCliente";
         }
         
         function redirectToEmpleado() {
-            window.location.href = "/validar";
+            window.location.href = "/validarEmpleado";
         }
     </script>
 </body>
