+++ src/main/resources/templates/registroEmpleado.html
@@ -79,7 +79,7 @@
 <body>
     <div class="container">
         <h2>Registro de empleado</h2>
-        <form action="/registro" method="post">
+        <form action="/registroEmpleado" method="post">
             <label for="nombre">Nombre:</label><br>
             <input type="text" id="nombre" name="nombre" required><br>
             <label for="email">Correo electrónico:</label><br>
